# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Syslogx API
  version: 0.4.0
  description: Syslogx ingestion, search, analytics, identity, export, and live-tail API.
servers:
  - url: /
paths:
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Login'}}}}
      responses: {'200': {description: Session created}, '401': {description: Invalid credentials}}
  /api/v1/auth/logout:
    post: {operationId: logout, responses: {'204': {description: Session removed}}}
  /api/v1/auth/me:
    get: {operationId: getCurrentUser, responses: {'200': {description: Current identity}, '401': {description: Authentication required}}}
  /api/v1/ingest:
    post:
      operationId: ingestJSON
      description: Accepts a JSON object, JSON array, or application/x-ndjson stream. Unknown fields are preserved.
      requestBody: {required: true, content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/IngestLog'}, {type: array, maxItems: 1000, items: {$ref: '#/components/schemas/IngestLog'}}]}}, application/x-ndjson: {schema: {type: string}}}}
      responses: {'202': {description: Batch accepted}, '400': {description: Invalid input}, '413': {description: Limit exceeded}}
  /api/v1/logs/search:
    post: {$ref: '#/components/pathItems/Query'}
  /api/v1/logs/stats:
    post: {$ref: '#/components/pathItems/Query'}
  /api/v1/fields:
    post: {$ref: '#/components/pathItems/Query'}
  /api/v1/fields/{field}/values:
    post:
      parameters: [{name: field, in: path, required: true, schema: {type: string}}]
      requestBody: {$ref: '#/components/requestBodies/Query'}
      responses: {'200': {description: Value counts}, '400': {description: Invalid query}}
  /api/v1/logs/export:
    post:
      parameters: [{name: format, in: query, schema: {type: string, enum: [json, csv, ndjson]}}]
      requestBody: {$ref: '#/components/requestBodies/Query'}
      responses: {'200': {description: Bounded streamed export}, '400': {description: Invalid query}}
  /api/v1/logs/tail:
    get:
      description: Server-Sent Events stream with a bounded server query and client buffer.
      parameters: [{name: query, in: query, schema: {type: string, maxLength: 4096}}]
      responses: {'200': {description: SSE event stream, content: {text/event-stream: {schema: {type: string}}}}}
  /api/v1/saved-searches:
    get: {responses: {'200': {description: Saved searches}}}
    post: {responses: {'201': {description: Saved search created}, '400': {description: Invalid input}}}
  /api/v1/saved-searches/{id}:
    delete:
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      responses: {'204': {description: Saved search deleted}, '404': {description: Search not found or not owned by caller}}
  /api/v1/sources:
    get:
      operationId: listSources
      description: Read-only configured listener list. Runtime source CRUD is not implemented.
      responses:
        '200': {description: Configured listeners, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Source'}}}}}}}
  /health:
    get:
      operationId: getHealth
      responses:
        '200': {description: Process is live}
  /ready:
    get:
      operationId: getReadiness
      responses:
        '200': {description: Required dependencies and ingestion are ready}
        '503': {description: One or more readiness checks failed}
  /metrics:
    get:
      operationId: getMetrics
      responses:
        '200': {description: Prometheus text exposition}
  /api/v1/capabilities:
    get:
      operationId: getCapabilities
      responses:
        '200':
          description: Storage capabilities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capabilities'
  /api/v1/system/logs/recent:
    get:
      operationId: getRecentLogs
      description: Phase 1 diagnostic endpoint limited to the last 15 minutes; not the final search API.
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 1000, default: 20}
      responses:
        '200': {description: Recent logs}
        '400': {description: Invalid limit}
        '502': {description: Storage unavailable}
components:
  pathItems:
    Query:
      requestBody: {$ref: '#/components/requestBodies/Query'}
      responses: {'200': {description: Query result}, '400': {description: Invalid query}}
  requestBodies:
    Query: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LogQuery'}}}}
  schemas:
    Source:
      type: object
      required: [id, name, protocol, address, parser, enabled, status]
      properties:
        id: {type: string}
        name: {type: string}
        protocol: {type: string}
        address: {type: string}
        parser: {type: string}
        enabled: {type: boolean}
    Login:
      type: object
      required: [username, password]
      properties: {username: {type: string}, password: {type: string, format: password}}
    IngestLog:
      type: object
      required: [message]
      properties: {timestamp: {type: string, format: date-time}, message: {type: string}, hostname: {type: string}, level: {type: string}, service: {type: string}}
      additionalProperties: true
    Filter:
      type: object
      required: [field, operator]
      properties: {field: {type: string}, operator: {type: string, enum: [eq, neq, contains, starts_with, exists, gt, gte, lt, lte]}, value: {}}
    LogQuery:
      type: object
      required: [start, end]
      properties:
        start: {type: string, format: date-time}
        end: {type: string, format: date-time}
        query: {type: string, maxLength: 4096}
        native_query: {type: string, maxLength: 8192}
        filters: {type: array, maxItems: 32, items: {$ref: '#/components/schemas/Filter'}}
        limit: {type: integer, minimum: 1, maximum: 1000, default: 100}
        cursor: {type: string}
    Capabilities:
      type: object
      required: [backend, full_text, field_discovery, facets, statistics, live_tail]
      properties:
        backend: {type: string}
        full_text: {type: boolean}
        field_discovery: {type: boolean}
        facets: {type: boolean}
        statistics: {type: boolean}
        live_tail: {type: boolean}
        native_dialect: {type: string}